3. report a say (it seems like that no matter on which say, i am clicking on, you always send the 1st say id and not the one i clicked on)
 4. profile menu click (i searched for a certain user, clicked to it and the profile view displayed. Great.

no, how do i go back to my profile? \
a click on the menu on the bottom should take me to my profile.always
make sure that, if I am already on the current profile that it is no re-loaded)

5. say colors - issue (I opened the add say dialog and the background colors was a randon color - great.
but, after I added the say the background color is completely different?