[25/02/2016, 2:47:44 PM] steve: 1. skip -  can see you implemented the skip functionality and that you are calling skipRateUserCharms
but, when you show a profile and rated="skip" at the profile level, you should show the charm rating of the profile for all charms. See above..

"if rated="skip" the current user can see the rating for the requested user (regardless of the rated value at the charm level)"
[25/02/2016, 2:49:10 PM] steve: 2. charms rating and reveal- (there is a small change in the rating process

today when a user did not rate a charm ("rating="false") he cannot see the rating and when he is rating he allowed to rate those charms...) --  I do not see that this is implemented.

I have a profile that I rated 1 of his charms and I still cannot see all the charms rating.